Handle the case when a gradient only have a single stop which contains missing components
In crrev.com/c/4834550 Aaron introduced
`ResolveStopColorWithMissingParams` for "take neighboring values for
gradient color stops with missing parameters", when I land
crrev.com/c/6197018, you can have only 1 single stop, if it also has
missing components, `ResolveStopColorWithMissingParams` will never be
called, and the result is we have an empty `colors` and `positions`
gradient, and crash the chrome. This CL added a branch to check if the
stops size is 1, and fixed the crash.

Ref: https://issues.chromium.org/issues/344620213#comment5
Bug: 344620213
Change-Id: Ia9c63b485c267f9b6b830357d2a1fdb1174fcb31